Remove Nose
Fixes #96
