+++ setup.cfg
@@ -1,7 +1,3 @@
-[nosetests]
-with-coverage=1
-cover-package=salmon
-
 [versioneer]
 VCS = git
 style = pep440
