+++ setup.py
@@ -38,8 +38,7 @@ config = {
     'install_requires': install_requires,
     'tests_require': tests_require,
     'extras_require': extras_require,
-    'setup_requires': ['nose'],
-    'test_suite': 'nose.collector',
+    'test_suite': 'tests',
     'packages': ['salmon', 'salmon.handlers'],
     'include_package_data': True,
     'name': 'salmon-mail',
