+++ tests/data/config/settings.py
@@ -1,0 +1,2 @@
+# Fake settings module for tests
+receiver_config = True
