+++ tests/data/templates/confirmation.msg
