+++ tests/data/test_app/README.md
@@ -1,0 +1,13 @@
+# test_app
+
+## What's this?
+
+It's a test app that is used by Salmon's integration tests
+
+## Can I use it?
+
+If you're writing tests, you might want to copy over some of the code here. If you're wanting to try out Salmon then this isn't for you.
+
+## I can't get the server started/accepting mail/etc.
+
+Unless that's causing Salmon's tests to fail, please don't report bugs against this app.
