+++ tests/data/test_app/config/testing.py
@@ -1,16 +1,13 @@
 import logging
 import logging.config
 
-import jinja2
-
-from salmon import view
 from salmon.routing import Router
 from salmon.server import Relay
 
 from . import settings
 
 # configure logging to go to a log file
-logging.config.fileConfig("tests/config/logging.conf")
+logging.config.fileConfig("tests/logging.conf")
 
 # the relay host to actually send the final message to (set debug=1 to see what
 # the relay is saying to the log server).
@@ -24,5 +21,3 @@ Router.defaults(**settings.router_defaults)
 Router.load(settings.handlers + settings.queue_handlers)
 Router.RELOAD = False
 Router.LOG_EXCEPTIONS = False
-
-view.LOADER = jinja2.Environment(loader=jinja2.PackageLoader('salmon_tests', 'templates'))
