--- tests/salmon_tests/templates/content.markdown
@@ -1,14 +1,0 @@
-Hello
-=====
-
-
-I would *love* for you to tell me what is going on here joe.  NOW!
-
-
-Alright
--------
-
-
-This is the best I can come up with.
-
-Zed
