--- tests/salmon_tests/templates/html_test.html
@@ -1,13 +1,0 @@
-<html>
-    <head>
-        <title>{{ title }}</title>
-    </head>
-
-    <body style="background: magenta">
-        <h1 class="bright">{{ title }}</h1>
-
-        {{ content }}
-
-        <h3 id="dull">All done.</h3>
-    </body>
-</html>
